# backend/openapi.yml
openapi: 3.0.3
info:
  title: Tickets API
  version: 1.0.0
  description: REST API для управления билетами, людьми, событиями и площадками.
servers:
  - url: /
...
paths:
...
  /add_venue:
    post:
      tags: [Venues]
      summary: Создать площадку
      operationId: addVenue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Venue' }
      responses:
        '200':
          description: Создано (data может содержать созданный Venue)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /get_venues:
    get:
      tags: [Venues]
      summary: Получить список площадок
      operationId: getVenues
      responses:
        '200':
          description: Успех (data = VenuesEnvelope)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
components:
  schemas:
    AbstractResponse:
      type: object
      properties:
        title: { type: string, nullable: true }
        message: { type: string, nullable: true }
        status: { type: string, nullable: true, example: "OK" }
        data: { nullable: true, description: "Произвольные данные ответа" }
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/AbstractResponse'
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/AbstractResponse'
...
    Venue:
      type: object
      properties:
        id: { type: integer, format: int32, nullable: true }
        name: { type: string }
        capacity: { type: integer, format: int32, nullable: true }
        type:
          type: string
          nullable: true
          enum: [ LOFT, OPEN_AREA, STADIUM ]
...